+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Change log
 
+## 1.1.0 11-Feb-2022
+
+- same type data now return `<type> | undefined`
+- export types
+
 ## 1.0.0 11-Feb-2022
 
 - 1st publish